requirements-windows.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt setup.py
#
aiohttp==2.3.10
aiohttp_cors==0.7.3
aiomysql==0.0.15
async-timeout==3.0.0
aiosmtplib==1.0.3
attrs==18.1.0             # via pytest
chardet==3.0.4
click==6.7
more-itertools==4.1.0     # via pytest
multidict==4.3.1
passlib==1.7.1
pluggy==0.6.0             # via pytest
py==1.5.3                 # via pytestf
pyjwt==1.6.4
pymysql==0.7.11
pytest-aiohttp==0.3.0
pytest==3.6.0            # via pytest-aiohttp
python-dotenv==0.7.1
#pyyaml==3.12
six==1.11.0               # via more-itertools, pytest
sqlalchemy==1.2.7
typing==3.6.4             # via aiohttp-cors
#uvloop==0.9.1
yarl==1.2.4
xlrd==1.1.0
pymongo==3.7.1
#boto==1.7.32
#cmd :sudo /usr/local/bin/python3 -m pip install -r requirements.txt
